+++ WebSalesMvcWithAngular/ClientApp/src/app/Components/SalesRecord/delete/delete-sales-record.component.ts
@@ -5,6 +5,7 @@ import { SalesRecordService } from '../../../Services/SalesRecordService';
 import { LoadingService } from '../../../Services/LoadingService';
 import { SaleStatus } from '../../../Models/enums/SaleStatus';
 import { PaymentMethod } from '../../../Models/enums/PaymentMethod';
+import { AlertService } from '../../../Services/AlertService';
 
 @Component({
   selector: 'app-delete-sales-record',
@@ -18,6 +19,7 @@ export class DeleteSalesRecordComponent implements OnInit {
     @Inject(MAT_DIALOG_DATA) public data: { salesRecord: SalesRecord },
     private salesRecordService: SalesRecordService,
     private loadingService: LoadingService,
+    private alertService: AlertService
   ) { }
 
   ngOnInit(): void {
@@ -27,14 +29,21 @@ export class DeleteSalesRecordComponent implements OnInit {
   }
 
   async onDeleteClick(): Promise<void> {
-
     try {
       if (this.salesRecord && this.salesRecord.id) {
-        await (await this.salesRecordService.deleteSalesRecord(this.salesRecord.id)).toPromise();
-        this.dialogRef.close({ deleted: true });
+        this.loadingService.showLoading();
+        if (this.salesRecord.status !== 1) {
+          await (await this.salesRecordService.deleteSalesRecord(this.salesRecord.id)).toPromise();
+          this.dialogRef.close({ deleted: true });
+        } else {
+          this.alertService.error('Vendas faturadas não podem ser deletadas, somente canceladas.');
+        }
       }
     } catch (error) {
       console.error('Erro ao deletar venda:', error);
+      this.alertService.error('Erro ao deletar venda:', error);
+    } finally {
+      this.loadingService.hideLoading();
     }
   }
